+++ src/main/java/Duke.java
@@ -1,8 +1,11 @@
+import duke.*;
+import duke.main.Parser;
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.Ui;
+import duke.command.Command;
+
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.NoSuchElementException;
 import java.util.Scanner;
 
 public class Duke {
@@ -47,7 +50,7 @@ public class Duke {
 
     }
 
-//    public static void getInput(Scanner scan) throws DukeException, IOException {
+//    public static void getInput(Scanner scan) throws duke.DukeException, IOException {
 //        String inData;
 //        try {
 //            inData = scan.nextLine();
@@ -63,14 +66,14 @@ public class Duke {
 //                ui.printExit();
 //                break;
 //            case("mark"):
-//                Task taskToMark = tasks.getTasks().get(Integer.parseInt(input[1]) - 1);
+//                duke.task.Task taskToMark = tasks.getTasks().get(Integer.parseInt(input[1]) - 1);
 //                taskToMark.markAsDone();
 //                ui.printMarkDone(taskToMark);
 //                storage.update(tasks);
 //                getInput(scan);
 //                break;
 //            case("unmark"):
-//                Task taskToUnmark = tasks.getTasks().get(Integer.parseInt(input[1]) - 1);
+//                duke.task.Task taskToUnmark = tasks.getTasks().get(Integer.parseInt(input[1]) - 1);
 //                taskToUnmark.markAsNotDone();
 //                ui.printMarkNotDone(taskToUnmark);
 //                storage.update(tasks);
@@ -82,27 +85,27 @@ public class Duke {
 //                break;
 //            case("todo"):
 //                    if (input.length == 1) {
-//                        throw new DukeException("The description of a todo cannot be empty.");
+//                        throw new duke.DukeException("The description of a todo cannot be empty.");
 //                    }
-//                    Todo curTodo = new Todo(input[1]);
+//                    duke.task.Todo curTodo = new duke.task.Todo(input[1]);
 //                    addToTasklist(curTodo);
 //                    getInput(scan);
 //                    break;
 //            case("deadline"):
 //                String[] deadlineString = input[1].split("/by ", 2);
 //                if (deadlineString.length == 1) {
-//                    throw new DukeException("The 'by date' of a deadline cannot be empty.");
+//                    throw new duke.DukeException("The 'by date' of a deadline cannot be empty.");
 //                }
-//                Deadline curDeadline = new Deadline(deadlineString[0], LocalDate.parse(deadlineString[1]));
+//                duke.task.Deadline curDeadline = new duke.task.Deadline(deadlineString[0], LocalDate.parse(deadlineString[1]));
 //                addToTasklist(curDeadline);
 //                getInput(scan);
 //                break;
 //            case("event"):
 //                String[]eventString = input[1].split("/at ", 2);
 //                if (eventString.length == 1) {
-//                    throw new DukeException("The 'at date' of a event cannot be empty.");
+//                    throw new duke.DukeException("The 'at date' of a event cannot be empty.");
 //                }
-//                Event curEvent = new Event(eventString[0], LocalDate.parse(eventString[1]));
+//                duke.task.Event curEvent = new duke.task.Event(eventString[0], LocalDate.parse(eventString[1]));
 //                addToTasklist(curEvent);
 //                getInput(scan);
 //                break;
@@ -111,13 +114,13 @@ public class Duke {
 //                getInput(scan);
 //                break;
 //            default:
-//                throw new DukeException("I'm sorry, but I don't know what that means :-(");
+//                throw new duke.DukeException("I'm sorry, but I don't know what that means :-(");
 //        }
 //
 //    }
 
 
-//    public static void addToTasklist(Task taskToAdd) throws IOException {
+//    public static void addToTasklist(duke.task.Task taskToAdd) throws IOException {
 //        tasks.addTask(taskToAdd);
 //        ui.printAdd(taskToAdd, tasks);
 //        storage.update(tasks);
