+++ src/main/java/duke/main/Parser.java
@@ -1,3 +1,11 @@
+package duke.main;
+
+import duke.DukeException;
+import duke.command.*;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Todo;
+
 import java.time.LocalDate;
 
 /**
@@ -7,7 +15,7 @@ public class Parser {
     /**
      * Make sense of user command input and return the respective command
      * @param userInput The line of command that user input.
-     * @return The respective Command.
+     * @return The respective duke.command.Command.
      * @throws DukeException If invalid command encountered.
      */
     public static Command parseCommand(String userInput) throws DukeException {
