+++ src/main/java/duke/task/Todo.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.DukeException;
+
 public class Todo extends Task {
 
     public Todo(String description) throws DukeException {
@@ -9,8 +13,8 @@ public class Todo extends Task {
 
     /**
      * Constructor for task in save file.
-     * @param description Description of the Todo.
-     * @param done If the Todo task is done or not done.
+     * @param description Description of the duke.task.Todo.
+     * @param done If the duke.task.Todo task is done or not done.
      */
     public Todo(String description, boolean done) {
         super(description, done);
